정수 반올림 되는 간단한 계산기 레이아웃_핑크

+++ app/src/main/java/com/example/calculator/MainActivity.java
@@ -19,52 +19,52 @@ public class MainActivity extends AppCompatActivity {
     }
 
 
-    // + 버튼 (덧셈)
+    // + 버튼 (덧셈 )
     public void onClickAdd(View view) {
         EditText number1 = (EditText) findViewById(R. id.editNum1);
         EditText number2 = (EditText) findViewById(R. id.editNum2);
         TextView result = (TextView) findViewById(R. id.result);
 
-        double n1 = Double.parseDouble(number1.getText().toString());
-        double n2 = Double.parseDouble(number2.getText().toString());
+        int n1 = Integer.parseInt(number1.getText().toString());
+        int n2 = Integer.parseInt(number2.getText().toString());
 
-        result.setText(Double.toString(n1 + n2));
+        result.setText(Integer.toString(n1 + n2));
     }
 
-    // - 버튼 (뺄셈)
+    // - 버튼 (뺄셈 )
     public void onClickSub(View view) {
         EditText number1 = (EditText) findViewById(R. id.editNum1);
         EditText number2 = (EditText) findViewById(R. id.editNum2);
         TextView result = (TextView) findViewById(R. id.result);
 
-        double n1 = Double.parseDouble(number1.getText().toString());
-        double n2 = Double.parseDouble(number2.getText().toString());
+        int n1 = Integer.parseInt(number1.getText().toString());
+        int n2 = Integer.parseInt(number2.getText().toString());
 
-        result.setText(Double.toString(n1 - n2));
+        result.setText(Integer.toString(n1 - n2));
     }
 
-    // * 버튼 (곱셈)
+    // * 버튼 (곱셈 )
     public void onClickMul(View view) {
         EditText number1 = (EditText) findViewById(R. id.editNum1);
         EditText number2 = (EditText) findViewById(R. id.editNum2);
         TextView result = (TextView) findViewById(R. id.result);
 
-        double n1 = Double.parseDouble(number1.getText().toString());
-        double n2 = Double.parseDouble(number2.getText().toString());
+        int n1 = Integer.parseInt(number1.getText().toString());
+        int n2 = Integer.parseInt(number2.getText().toString());
 
-        result.setText(Double.toString(n1 * n2));
+        result.setText(Integer.toString(n1 * n2));
     }
 
-    // / 버튼 (나눗셈)
+    // / 버튼 (나눗셈 )
     public void onClickDiv(View view) {
         EditText number1 = (EditText) findViewById(R. id.editNum1);
         EditText number2 = (EditText) findViewById(R. id.editNum2);
         TextView result = (TextView) findViewById(R. id.result);
 
-        double n1 = Double.parseDouble(number1.getText().toString());
-        double n2 = Double.parseDouble(number2.getText().toString());
+        int n1 = Integer.parseInt(number1.getText().toString());
+        int n2 = Integer.parseInt(number2.getText().toString());
 
-        result.setText(Double.toString(n1 / n2));
+        result.setText(Integer.toString(n1 / n2));
     }
 
 }
